pyproject.toml
[project]
name = "simple-investment-portfolio"
version = "0.1.0"
description = "Демонстрационный проект, написанный на Django с использованием чистой архитектуры и DDD"
requires-python = ">=3.13"
dependencies = [
    "dishka>=1.6.0",
    "django>=5.2.3",
    "djangorestframework>=3.16.0",
    "drf-yasg>=1.21.10",
]

[dependency-groups]
dev = [
    "mypy>=1.16.1",
    "pre-commit>=4.2.0",
    "ruff>=0.12.1",
]

[tool.mypy]
disallow_any_generics = true
disallow_subclassing_any = false
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
strict_equality = true
disallow_untyped_decorators = false
ignore_missing_imports = true
implicit_reexport = true

[tool.ruff]
src = ["src"]
line-length = 80

[tool.ruff.lint]
select = [
    # isort
    "I001"
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
